requirements.txt
llama-index-core>=0.10.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
anthropic>=0.39.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.20
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pypdf>=4.0.0
python-docx>=1.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
